requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-chess==1.999
requests==2.31.0
python-dotenv==1.0.0
